Added tags to the title for common failure conditions
There are two common failure modes for e(x)hentai galleries
  1) the gallery is expunged (eg copyright takedown)
  2) the user is out of GP to download the archive
Failure 2 is easily detectable as the response to getting the
archive URL specifically reports such a failure.
Failure 1 is more of an educated guess as the point of detection
is getting a 404 even though we were given an archive key.

Also a seemingly innocuous bug of the condition of which login
service to use. The condition to choose x or non-x was backwards.