Failed attempt at expunged galleries
This is only a failure because the flag seems to make zero sense.
I have examples where things are marked as expunged but are still
totally available on both domains, archives included.
I also have examples where the archives are not available but the
expunged flag is still false.
I have not found a single example yet where a gallery is marked as
expunged and is unavailable, ergo, the flag is useless.