Small refactor to the ehentai downloader
The code was a bit silly with the indentation due to nesting things
inside of if statements that never even had an else clause.
Mostly code was just moved up in scope and the if statements were
inverted to return early when meeting the pre-condition.

Also added in comments for different sections and organized the sections
to keep logical parts together.